+++ spec/models/answer_spec.rb
@@ -7,4 +7,8 @@ RSpec.describe Answer, type: :model do
     answer = build(:answer, response: "Yellow")
     expect(answer.response).to eql("Yellow")
   end
+
+  describe "validations" do
+    it { is_expected.to validate_presence_of(:response) }
+  end
 end